backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==1.63.0
azure-identity==1.17.1
azure-core==1.30.2
pydantic==2.10.3
python-multipart==0.0.20
httpx==0.27.0
python-dotenv==1.0.1
supabase==2.10.0
postgrest==0.17.1
gotrue==2.10.1
realtime==1.0.6
storage3==0.8.1
